renamed config package to state
