+++ cli/src/cmd/root.go
@@ -7,7 +7,7 @@ import (
 	"github.com/spf13/cobra"
 
 	"cli/cmd/commands"
-	"cli/core/config"
+	"cli/core/state"
 )
 
 // rootCmd represents the base command when called without any subcommands
@@ -30,9 +30,9 @@ func init() {
 	// TODO: read the docs for cobra.OnInitialize() and decide if it's needed
 	cobra.OnInitialize()
 
-	rootCmd.PersistentFlags().StringVar(&config.ConfigFile, "config", "", "config file (default is $WORKING_DIR/config.yaml)")
+	rootCmd.PersistentFlags().StringVar(&state.ConfigFile, "config", "", "config file (default is $WORKING_DIR/config.yaml)")
 	// Note: No shorthand for env-file, saving -e for individual env var declarations
-	rootCmd.PersistentFlags().StringSliceVar(&config.EnvFiles, "env-file", nil, "env file")
+	rootCmd.PersistentFlags().StringSliceVar(&state.EnvFiles, "env-file", nil, "env file")
 
 	commands.AddCommands(rootCmd)
 }
