+++ cli/src/core/state/config.go
@@ -1,4 +1,4 @@
-package config
+package state
 
 var (
 	ConfigFile string
